[core] Modularize batchnorm. (#2996)
