+++ tfjs-core/src/public/chained_ops/register_all_chained_ops.ts
@@ -23,3 +23,4 @@ import './tile';
 import './one_hot';
 import './transpose';
 import './pad';
+import './batchnorm';
