+++ tfjs-core/src/register_all_gradients.ts
@@ -16,6 +16,7 @@
  */
 import {broadcastToGradConfig} from './gradients/BroadcastTo_grad';
 import {divGradConfig} from './gradients/Div_grad';
+import {fusedBatchNormGradConfig} from './gradients/FusedBatchNorm_grad';
 import {identityGradConfig} from './gradients/Identity_grad';
 import {oneHotGradConfig} from './gradients/OneHot_grad';
 import {padV2GradConfig} from './gradients/PadV2_grad';
@@ -28,9 +29,9 @@ import {registerGradient} from './kernel_registry';
 
 // Export all kernel configs here so that the package can auto register them
 const gradConfigs: GradConfig[] = [
-  divGradConfig, squareGradConfig, squaredDifferenceGradConfig,
-  broadcastToGradConfig, identityGradConfig, tileGradConfig, oneHotGradConfig,
-  transposeGradConfig, padV2GradConfig
+  broadcastToGradConfig, divGradConfig, fusedBatchNormGradConfig,
+  identityGradConfig, oneHotGradConfig, padV2GradConfig, squareGradConfig,
+  squaredDifferenceGradConfig, tileGradConfig, transposeGradConfig
 ];
 
 for (const gradientConfig of gradConfigs) {
